wire v0x03: source_id + target_group widened to u16 LE
Bumps kProtocolVersion 0x02 -> 0x03. Header source_id and every
payload's target_group widen from 1 byte to 2 bytes little-endian.

Motivation is futureproofing:
- source_id 191 slots -> 65534 slots. Birthday-paradox says 50%
  collision at ~14 concurrent Directors on v2 (191 slots), enough
  to hit at a single dense personal-rig stage. v3 pushes that
  ceiling well past any realistic crowd.
- target_group 255 slots -> 65534 slots. Stadium-scale seat-level
  addressing headroom - not needed by current fx but locks in the
  wire so a future addressing scheme doesn't need a second break.

DAL / config / NVS all stay at u8. UI unchanged. The u16 width is
purely wire-side; top byte is zero for current-generation Directors,
and a later NVS/UI widening slots in without another wire break.

Header layout (v3): magic(2) + version(1) + source_id(2 LE) + seq(1)
+ hop(1) + type(1) + payload_len(1) = 9 bytes (v2 was 8).
kHopCountOffset 5 -> 6. Every LIGHT_* / TEXT / BITMAP / CLEAR payload
grows +1 byte. LIGHT_PULSE 9 -> 10, WASH 16 -> 17, TEXT prefix 6 -> 7,
BITMAP_HEADER 37 -> 38, BITMAP_PLANE prefix 5 -> 6, CLEAR 3 -> 4.

Old-firmware devices reject v3 at the version check; this is a hard
cutover, not a compatible extension. Coordinated with the tildagon-
firmware v0x03 PR - both merge in one bench session to avoid mixed-
version network state.

Test coverage: 36/36 native_espnow (existing byte-for-byte fixtures
regenerated + new v3 wide-range coverage), 11/11 native_modes,
56/56 native_dal, 14/14 native_effects, 15/15 native_audio. All four
device firmwares build clean (Plus2, S3, Atom Lite, AtomS3R-PoE).

Tildagon firmware + docs land in separate PRs on the same cutover.

+++ src/modes/repeater_mode.cpp
@@ -292,7 +292,7 @@ void RepeaterMode::update_led(uint32_t now) {
 #endif
 }
 
-bool RepeaterMode::seen_recently(uint8_t src, uint8_t seq) const {
+bool RepeaterMode::seen_recently(uint16_t src, uint8_t seq) const {
     if (seq == 0) return false;   // seq 0 = sequencing disabled, never dedup
     for (size_t i = 0; i < kDedupRingSize; ++i) {
         if (dedup_ring_[i].source_id == src &&
@@ -303,7 +303,7 @@ bool RepeaterMode::seen_recently(uint8_t src, uint8_t seq) const {
     return false;
 }
 
-void RepeaterMode::mark_seen(uint8_t src, uint8_t seq) {
+void RepeaterMode::mark_seen(uint16_t src, uint8_t seq) {
     if (seq == 0) return;
     dedup_ring_[dedup_head_] = DedupEntry{src, seq};
     dedup_head_ = (dedup_head_ + 1) % kDedupRingSize;
